+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Filter
 * Table Ajax
 
+## Improve Tabs Component
+
+* Passing a prop of align='right' will align tab headers to the right
+
 ## Bug Fixes
 
 * TableCell and TableHeader can receive additional props.